requirements.txt
# ─── Core Framework ──────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
python-dotenv>=1.0.1
python-multipart>=0.0.9

# ─── Database ────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# ─── Auth ────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# ─── Validation ──────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.0

# ─── ML / NLP ────────────────────────────────────────────────
transformers>=4.38.0
torch>=2.2.0
sentence-transformers>=2.5.0
scikit-learn>=1.4.0
nltk>=3.8.0
spacy>=3.7.0

# ─── RAG Pipeline ────────────────────────────────────────────
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0

# ─── Document Extraction ─────────────────────────────────────
pymupdf>=1.23.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
chardet>=5.2.0

# ─── HTTP / Utilities ────────────────────────────────────────
httpx>=0.27.0
aiofiles>=23.2.0

# ─── Monitoring (optional) ───────────────────────────────────
psutil>=5.9.0
